[tests/modules] Moved all test case definitions to modules. Renamed modules.py to entities.py
